Making progress and necessary edits to maintain consistency in typing; making progress on the PatientRecordService that returns all of the information necessary for a patientRecord

+++ UnifiedPatientManager/src/main/java/com/Eges411Team/UnifiedPatientManager/repositories/AllergyRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface AllergyRepository extends JpaRepository<Allergy, Integer> {
 
     // Uses the Java field name "patient_id"
-    List<Allergy> findAllByPatient_id(int patient_id);
+    List<Allergy> findAllByPatient_id(Long patient_id);
 }
